requirements.txt
streamlit
requests
pandas
altair
python-dateutil
